third_party/blink/web_tests/external/wpt/css/css-borders/tentative/parsing/box-shadow-position-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property box-shadow-position value 'inset'
  assert_true: box-shadow-position doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-position value 'outset'
  assert_true: box-shadow-position doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-position value 'inset, outset'
  assert_true: box-shadow-position doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-position value 'outset, inset'
  assert_true: box-shadow-position doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
